fix: update expected mod counts and enhance error handling in AvailableModsPanel
- Updated expected mod counts for "claw" and "staff" in test_mod_counts.py to reflect correct values.
- Added error state management in AvailableModsPanel to display error messages when loading mods fails.
- Modified ItemCreationPanel to include additional weapon categories in the preferred order and added unrevealed_mods to the item structure.

+++ backend/app/schemas/crafting.py
@@ -79,8 +79,8 @@ class CraftableItem(BaseModel):
     unrevealed_mods: List[UnrevealedModifier] = Field(default_factory=list)  # Unrevealed desecrated mods
 
     corrupted: bool = False
-    base_stats: Dict[str, int] = Field(default_factory=dict)  # Base stats from item base
-    calculated_stats: Dict[str, int] = Field(default_factory=dict)  # Final calculated stats
+    base_stats: Dict[str, float] = Field(default_factory=dict)  # Base stats from item base (int or float)
+    calculated_stats: Dict[str, float] = Field(default_factory=dict)  # Final calculated stats (int or float)
 
     @property
     def prefix_count(self) -> int:
